+++ PyPlayerokAPI/stream/listener/__init__.py
@@ -4,7 +4,7 @@ from .chat_storage import ChatStorage
 from .deals_watcher import DealWatcher
 from .message_resolver import MessageResolver
 from .review_watcher import ReviewWatcher
-from .websocket_client import WebsocketClient, subprotocols
+from .websocket_client import WebsocketClient
 
 
 __all__ = [
@@ -13,5 +13,4 @@ __all__ = [
     "MessageResolver",
     "ReviewWatcher",
     "WebsocketClient",
-    "subprotocols",
 ]